refactor, add doc strings and description for module
